Put the same code into `BaseFileChecksumHelper`
