+++ hadoop-ozone/client/src/test/java/org/apache/hadoop/ozone/client/checksum/TestECFileChecksumHelper.java
@@ -1,0 +1,4 @@
+package org.apache.hadoop.ozone.client.checksum;
+
+public class TestECFileChecksumHelper {
+}
